sessions/01_file_structure_classification/requirements.txt
google-cloud-bigquery[bqstorage,pandas]
google-auth
polars
pyarrow
pandas-gbq